+++ XFaction/CHANGELOG.md
@@ -2,6 +2,7 @@
 [XFaction]
 - Local guild logouts and gchat are now in XFaction format
 - Configurable team names (Zoombara)
+- Respects guild rank gchat speak/listen permissions
 
 ### Version 3.3.3 [ 07.17.2022 ]
 [XFaction]